% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis.R
\name{bis_property_prices}
\alias{bis_property_prices}
\title{BIS residential property price indices}
\usage{
bis_property_prices(country = "US", last_n = 60)
}
\arguments{
\item{country}{2-letter country code}

\item{last_n}{Last N observations (default 60)}
}
\value{
tibble: date, value, ref_area, title
}
\description{
BIS residential property price indices
}
